% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{as.pdf}
\alias{as.pdf}
\title{Coerce to PDF}
\usage{
as.pdf(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Coerces to PDF. Generic, with default method.
}
\examples{
example(as.pdf.default)
}
\seealso{
Other as.pdf: \code{\link{as.pdf.default}}

Other devices: \code{\link{as.pdf.default}},
  \code{\link{as.png.default}}, \code{\link{as.png}}
}
