% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{as.png}
\alias{as.png}
\title{Coerce to PNG}
\usage{
as.png(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Coerces to PNG. Generic, with default method.
}
\examples{
example(as.png.default)
}
\seealso{
Other as.png: \code{\link{as.png.default}}

Other devices: \code{\link{as.pdf.default}},
  \code{\link{as.pdf}}, \code{\link{as.png.default}}
}
