% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{as.size.list}
\alias{as.size.list}
\title{Coerce List to Size}
\usage{
\method{as.size}{list}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\description{
Coerce list to class 'size'.
}
\seealso{
Other size: \code{\link{as.size}}
}
