% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{as.size}
\alias{as.size}
\title{Coerce to Size}
\usage{
as.size(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\description{
Coerce to class 'size'.  Generic, with method for list.
}
\seealso{
Other generic functions: \code{\link{devsize}},
  \code{\link{plotsize}}, \code{\link{unscaled}}

Other devsize: \code{\link{devsize.default}},
  \code{\link{devsize}}, \code{\link{print.size}},
  \code{\link{unscaled.default}}, \code{\link{unscaled}}

Other plotsize: \code{\link{plotsize.default}},
  \code{\link{plotsize}}, \code{\link{print.size}}

Other size: \code{\link{as.size.list}}
}
