% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{devsize}
\alias{devsize}
\title{Determine Device Size}
\usage{
devsize(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments}
}
\description{
Determines the size of a device.  Generic, with default method.
}
\examples{
example(devsize.default)
}
\seealso{
Other generic functions: \code{\link{as.size}},
  \code{\link{plotsize}}, \code{\link{unscaled}}

Other devsize: \code{\link{as.size}},
  \code{\link{devsize.default}}, \code{\link{print.size}},
  \code{\link{unscaled.default}}, \code{\link{unscaled}}
}
