% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{plotsize}
\alias{plotsize}
\title{Determine Plot Size}
\usage{
plotsize(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments}
}
\description{
Determines the size of a plot.  Generic, with default method.
}
\examples{
example(plotsize.default)
}
\seealso{
Other generic functions: \code{\link{as.size}},
  \code{\link{devsize}}, \code{\link{unscaled}}

Other plotsize: \code{\link{as.size}},
  \code{\link{plotsize.default}}, \code{\link{print.size}}
}
