% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{print.size}
\alias{print.size}
\title{Print Size}
\usage{
\method{print}{size}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\description{
Print object of class 'size', as returned by \code{\link{devsize}} and \code{\link{plotsize}}.
}
\seealso{
Other plotsize: \code{\link{as.size}},
  \code{\link{plotsize.default}}, \code{\link{plotsize}}

Other devsize: \code{\link{as.size}},
  \code{\link{devsize.default}}, \code{\link{devsize}},
  \code{\link{unscaled.default}}, \code{\link{unscaled}}
}
