% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{unscaled.default}
\alias{unscaled.default}
\title{Determine Unscaled Device Size by Default}
\usage{
\method{unscaled}{default}(x, width, height, ...)
}
\arguments{
\item{x}{object}

\item{width}{desired width in inches}

\item{height}{desired height in inches}

\item{...}{other arguments}
}
\description{
Determines the size of an unscaled device. Width and height
are interpreted directly as device dimensions rather than plot dimensions.
}
\seealso{
Other devsize: \code{\link{as.size}},
  \code{\link{devsize.default}}, \code{\link{devsize}},
  \code{\link{print.size}}, \code{\link{unscaled}}
}
