% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{unscaled}
\alias{unscaled}
\title{Determine Unscaled Device Size}
\usage{
unscaled(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments}
}
\description{
Determines the size of an unscaled device.  Generic, with default method.
}
\examples{
example(plotsize.default)
}
\seealso{
Other generic functions: \code{\link{as.size}},
  \code{\link{devsize}}, \code{\link{plotsize}}

Other devsize: \code{\link{as.size}},
  \code{\link{devsize.default}}, \code{\link{devsize}},
  \code{\link{print.size}}, \code{\link{unscaled.default}}
}
